Rename and move Locale to l10n::L10n
